requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
django-filter==24.2